requirements.txt
pandas==1.3.1
requests==2.25.1
psycopg2_binary==2.9.6
pytest-mock==3.11.1
